+++ CHANGELOG.md
@@ -4,8 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-## [1.1.2] - 2016-08-09
+## [1.2.0] - 2016-08-10
+### Added
+* If the template argument is a string name, and it doesn't exist, we
+  warn the user to ensure the `templating` packing is installed.
+* Ability to pass a Blaze.Template instance directly to the component's
+  template arg, and if neither that nor a string is provided, throw a
+  clear error.
 
+## [1.1.2] - 2016-08-09
 ### Fixed
 * Require `templating` as a direct package dep instead of assuming the user is
   using templates in their app, fixes #5.
@@ -26,7 +33,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial release.
 
-[Unreleased]: https://github.com/gadicc/meteor-blaze-react-component/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/gadicc/meteor-blaze-react-component/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/gadicc/meteor-blaze-react-component/compare/v1.1.2...v1.1.0
 [1.1.2]: https://github.com/gadicc/meteor-blaze-react-component/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/gadicc/meteor-blaze-react-component/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/gadicc/meteor-blaze-react-component/compare/v1.0.0...v1.1.0